Move waitForDeviceOnline method to RokuDevice
